third_party/WebKit/LayoutTests/inspector/profiler/heap-profiler-profiling-expected.txt

Tests that sampling heap profiling works. Bug 52634.

Profiler was enabled.

Running: testProfiling
found pageFunction heap-profiler-profiling.html:7
found (anonymous function) heap-profiler-profiling.html:8

Profiler was disabled.
